+++ scenario-compiler/src/test/scala/pl/touk/nussknacker/engine/spel/SpelExpressionSpec.scala
@@ -28,10 +28,6 @@ import pl.touk.nussknacker.engine.api.{Context, NodeId, SpelExpressionExcludeLis
 import pl.touk.nussknacker.engine.definition.clazz.{ClassDefinitionSet, JavaClassWithVarargs}
 import pl.touk.nussknacker.engine.dict.SimpleDictRegistry
 import pl.touk.nussknacker.engine.expression.parse.{CompiledExpression, TypedExpression}
-import pl.touk.nussknacker.engine.spel.SpelExpressionParseError.ArrayConstructorError.{
-  EmptyArrayDimension,
-  IllegalArrayDimensionType
-}
 import pl.touk.nussknacker.engine.spel.SpelExpressionParseError.IllegalOperationError.{
   InvalidMethodReference,
   TypeReferenceError
@@ -42,6 +38,7 @@ import pl.touk.nussknacker.engine.spel.SpelExpressionParseError.MissingObjectErr
   UnknownMethodError
 }
 import pl.touk.nussknacker.engine.spel.SpelExpressionParseError.OperatorError._
+import pl.touk.nussknacker.engine.spel.SpelExpressionParseError.UnsupportedOperationError.ArrayConstructorError
 import pl.touk.nussknacker.engine.spel.SpelExpressionParseError.{ArgumentTypeError, ExpressionTypeError}
 import pl.touk.nussknacker.engine.spel.SpelExpressionParser.{Flavour, Standard}
 import pl.touk.nussknacker.engine.testing.ModelDefinitionBuilder
@@ -1250,39 +1247,9 @@ class SpelExpressionSpec extends AnyFunSuite with Matchers with ValidatedValuesD
     }
   }
 
-  test("should type and evaluate empty array from array constructor") {
-    val result = parse[Any]("new String[0]", ctx)
-    result.validValue.returnType shouldBe Typed.typedClass[Array[String]]
-    result.validExpression.evaluateSync[Array[String]](ctx) shouldBe Array[String]()
-  }
-
-  test("should type and evaluate empty array from array constructor with values") {
-    val result = parse[Any]("new String[2]{'str1', 'str2'}", ctx)
-    result.validValue.returnType shouldBe Typed.fromInstance(Array("str1", "str2"))
-    result.validExpression.evaluateSync[Array[String]](ctx) shouldBe Array("str1", "str2")
-  }
-
-  test("should type and evaluate empty multidimensional array from array constructor") {
-    val result = parse[Any]("new String[0][0]", ctx)
-    result.validValue.returnType shouldBe Typed.typedClass[Array[Array[String]]]
-    result.validExpression.evaluateSync[Array[String]](ctx) shouldBe Array[Array[String]]()
-  }
-
-  test("should not validate empty dimensions in array constructor") {
-    List("new String[]", "new String[ ]").map(illegalExpr =>
-      parse[Any](illegalExpr, ctx).invalidValue shouldBe NonEmptyList.one(EmptyArrayDimension)
-    )
-  }
-
-  test("should not validate non-numeric type in dimension of array constructor") {
-    parse[Any]("new String['str']", ctx).invalidValue shouldBe NonEmptyList.one(
-      IllegalArrayDimensionType(Typed.fromInstance("str"))
-    )
-  }
-
-  test("should not allow accessing methods on unknown in dimensions of array constructor") {
-    parse[Any]("new String[{1,2,3}.remove(1) == null ? 0 : 0]", ctx).invalidValue shouldBe NonEmptyList.one(
-      UnknownMethodError("remove", "List[Unknown]")
+  test("should not validate array constructor") {
+    List("new String[]", "new String[ ]", "new String[0]", "new String[#invalidRef]", "new String[invalidSyntax]").map(
+      illegalExpr => parse[Any](illegalExpr, ctx).invalidValue shouldBe NonEmptyList.one(ArrayConstructorError)
     )
   }