+++ docs/Changelog.md
@@ -12,9 +12,9 @@
 * [#6388](https://github.com/TouK/nussknacker/pull/6388) Fix issue with suggestion expression mode and any value with suggestion in fragmentInput component, now supporting SpEL expressions.
 * [#6269](https://github.com/TouK/nussknacker/pull/6269) SpEL expression validator now works with values that are lists or maps.
   * NOTE: selection (`.?`), projection (`.!`) or operations from the `#COLLECTIONS` helper aren't properly supported in validation expressions, and will cause the result to always be invalid
-* [#6386](https://github.com/TouK/nussknacker/pull/6386) Security fix: removed ability to execute unallowed expressions
-  inside constructor array dimension definition
-* [#6386](https://github.com/TouK/nussknacker/pull/6386) Fixed typing of arrays
+* [#6386](https://github.com/TouK/nussknacker/pull/6386) Security fix for situation where array constructor could be 
+  used to execute unallowed expressions by exploiting the lack of full validation inside array dimensions. 
+  * Array constructor expressions are now illegal.
 
 1.16.1 (16 July 2024)
 -------------------------